metrics: fixes to Precision-Recall curve
The pixel classification metrics were originally tested with smoothly distributed values. For example values output by a softmax would range from 0 to 1 and given a usual U-Net we would observe smoothly changing logit levels.

The implementation however creates wrong results in some cases - for instance if we the prediction is binary, score = 0 or 1.

    In dynamic binning mode, set bin edges to surround the whole range of values by adding extra bin edges at minval - eps and maxval + eps. This allows us to correctly treat groups of sharp 0s and 1s scores.

    When calculating PR curve, ensure the curve starts at precision=1,recall=0 and ends at recall=1

    Area under PR curve is calculated according to sklearn and ignores the areas where the curve rises, instead taking a flat section of the graph. (see images)

There is still some mismatch when there are lots of repeated thresholds.